Add S3Service.
