+++ src/main/java/se/nbis/lega/inbox/s3/S3Service.java
@@ -1,0 +1,31 @@
+package se.nbis.lega.inbox.s3;
+
+import com.amazonaws.services.s3.AmazonS3;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@ConditionalOnBean(AmazonS3.class)
+@Service
+public class S3Service {
+
+    private AmazonS3 amazonS3;
+
+    public void prepareBucket(String username) {
+        try {
+            if (!amazonS3.doesBucketExistV2(username)) {
+                amazonS3.createBucket(username);
+            }
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
+        }
+    }
+
+    @Autowired
+    public void setAmazonS3(AmazonS3 amazonS3) {
+        this.amazonS3 = amazonS3;
+    }
+
+}
